py-blog-tool/ui_tool/requirements.txt
# Vue-Blog UI Tool Requirements

# Core dependencies
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# Data visualization
matplotlib==3.8.0
pandas==2.1.0

# Markdown processing
markdown==3.4.0

# Configuration handling
pyyaml==6.0.1

# Character encoding detection
chardet==5.2.0

# File system operations
pathlib==1.0.1

# Logging
logging==0.5.1.2

# Time and date handling
datetime==5.0

# JSON handling
json5==0.9.14

# OS operations
os-sys==2.1.4